separate storage utils
