+++ client/src/utils/index.ts
@@ -1,3 +1,4 @@
 export { default as setRef } from './setRef';
 export * from './format';
 export * as DB from './DB';
+export * from './storage';
